simulator/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
